+++ src/components/status_tree.rs
@@ -419,6 +419,14 @@ impl Component for StatusTreeComponent {
 			)
 			.order(order::RARE_ACTION),
 		);
+		out.push(
+			CommandInfo::new(
+				strings::commands::edit_item(&self.key_config),
+				self.selection_file().is_some(),
+				self.focused || force_all,
+			)
+			.order(order::RARE_ACTION),
+		);
 
 		CommandBlocking::PassingOn
 	}
@@ -456,6 +464,18 @@ impl Component for StatusTreeComponent {
 									),
 								),
 							));
+						}
+					}
+					Ok(EventState::Consumed)
+				} else if key_match(e, self.key_config.keys.edit_file)
+				{
+					if let Some(status_item) = self.selection_file() {
+						if let Some(queue) = &self.queue {
+							queue.push(
+								InternalEvent::OpenExternalEditor(
+									Some(status_item.path),
+								),
+							);
 						}
 					}
 					Ok(EventState::Consumed)